Fix plotting in the limit.
When plotting the 1stddev range over a value, we got into trouble if it
is virtually non-existent. Fix this by recognizing the case and then
avoid plotting the stddev range. Later code will add the A/M and this
will be in the right spot. Had we put anything into the plot it would
have been overwritten anyway.